requirements.txt
dash
dash-bootstrap-components
newspaper3k
openai
httpx
python-dotenv